Try uploading this code coverage summary instead

+++ Source/GridBased2D/Private/Characters/GameCharacter.cpp
@@ -17,7 +17,7 @@ AGameCharacter::AGameCharacter() {
 	// Set this actor to call Tick() every frame.  You can turn this off to improve performance if you don't need it.
 	PrimaryActorTick.bCanEverTick = true;
 
-	float BoxSize = GridBased2D::GRID_SIZE / 2;
+	float BoxSize = static_cast<float>(GridBased2D::GRID_SIZE) / 2;
 	auto Capsule = GetCapsuleComponent();
 	Capsule->SetCapsuleRadius(BoxSize);
 	Capsule->SetCapsuleHalfHeight(BoxSize);
